% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulation}
\alias{simulation}
\title{Simulate a dataset which can be analyzed by SLTCA}
\usage{
simulation(n)
}
\arguments{
\item{n}{Sample size.}
}
\value{
Returns a data frame with 6 longitudinal features y.1 - y.6, including count (y.1 and y.2), binary (y.3 and y.4) and continuous (y.5 and y.6) type. Variable baselinecov is the baseline risk factor of latent classes. Variable latent is the true latent class labels.
}
\description{
Simulate a dataset with longitudinal observations.
}
\examples{

dat <- simulation(500)
}
\author{
Teng Fei. Email: tfei@emory.edu
}
