% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLTCA.r
\name{SLTCA}
\alias{SLTCA}
\title{SLTCA: Scalable and Robust Latent Trajectory Class Analysis Using Artificial Likelihood}
\usage{
SLTCA(k = 20, dat, num_class, id, time, num_obs, features, Y_dist, covx, ipw,
  stop, tol = 0.005, max = 50, varest = T, MSC = "EQIC", verbose = T)
}
\arguments{
\item{k}{Number of random initialization to start the algorithm.}

\item{dat}{Input data matrix.}

\item{num_class}{Number of latent classes in the fitted model.}

\item{id}{Column name in the data matrix `dat` for the patient id.}

\item{time}{Column name in the data matrix `dat` for the time of longitudinal observations.}

\item{num_obs}{Column name in the data matrix `dat` for the number of longitudinal observations (number of visits).}

\item{features}{A vector of column names in the data matrix `dat` for the longitudinal observations.}

\item{Y_dist}{A vector indicating the type of longitudinal observations. An element of Y_dist can be 'normal','bin', and 'poi' for continuous, binary and count data.}

\item{covx}{A vector of column names in the data matrix `dat` for baseline latent class risk factors.}

\item{ipw}{Column name in the data matrix `dat` for the inverse probability weights for missingness. ipw=1 if not specified.}

\item{stop}{Stopping criterion for the algorithm. stop can be either 'tau' based on posterior probabilities or 'par' based on point estimation.}

\item{tol}{A constant such that the algorithm stops if the stopping criterion is below this constant.}

\item{max}{Maximum number of iterations if the algorithm does not converge.}

\item{varest}{True or False: whether conduct variance estimation or not.}

\item{MSC}{Model selection criteria: 'AQIC','BQIC' or 'EQIC'.}

\item{verbose}{Output progress of fitting the model.}
}
\description{
Conduct latent trajectory class analysis with longitudinal observations.
}
\author{
Teng Fei. Email: tfei@emory.edu
}
